# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AppSettingsOut": {
        "properties": {
          "approval_threshold": {
            "title": "Approval Threshold",
            "type": "number"
          },
          "tax_rate": {
            "title": "Tax Rate",
            "type": "number"
          }
        },
        "required": [
          "approval_threshold",
          "tax_rate"
        ],
        "title": "AppSettingsOut",
        "type": "object"
      },
      "AppSettingsUpdate": {
        "additionalProperties": false,
        "properties": {
          "approval_threshold": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,15}0*$)\\d{0,12}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Threshold"
          },
          "tax_rate": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,3}|(?=[\\d.]{1,6}0*$)\\d{0,3}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Rate"
          }
        },
        "title": "AppSettingsUpdate",
        "type": "object"
      },
...
      "ApprovalStats": {
        "properties": {
          "approved": {
            "title": "Approved",
            "type": "integer"
          },
          "average_decision_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Decision Hours"
          },
          "cancelled": {
            "title": "Cancelled",
            "type": "integer"
          },
          "pending": {
            "title": "Pending",
            "type": "integer"
          },
          "pending_value": {
            "title": "Pending Value",
            "type": "number"
          },
          "rejected": {
            "title": "Rejected",
            "type": "integer"
          }
        },
        "required": [
          "pending",
          "approved",
          "rejected",
          "cancelled",
          "pending_value",
          "average_decision_hours"
        ],
        "title": "ApprovalStats",
        "type": "object"
      },
...
      "CustomerCreate": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "name": {
            "maxLength": 150,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "pattern": "^[0-9+()\\-\\s]{7,20}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "required": [
          "name",
          "email"
        ],
        "title": "CustomerCreate",
        "type": "object"
      },
      "CustomerOut": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "address",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "CustomerOut",
        "type": "object"
      },
...
      "CustomerUpdate": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 150,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "pattern": "^[0-9+()\\-\\s]{7,20}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "title": "CustomerUpdate",
        "type": "object"
      },
      "DashboardSummary": {
        "properties": {
          "approvals": {
            "$ref": "#/components/schemas/ApprovalStats"
          },
          "customers": {
            "title": "Customers",
            "type": "integer"
          },
          "inventory": {
            "$ref": "#/components/schemas/InventoryStats"
          },
          "orders": {
            "$ref": "#/components/schemas/OrderCounts"
          },
          "sales": {
            "$ref": "#/components/schemas/SalesSummary"
          }
        },
        "required": [
          "sales",
          "orders",
          "approvals",
          "inventory",
          "customers"
        ],
        "title": "DashboardSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InventoryStats": {
        "properties": {
          "active_products": {
            "title": "Active Products",
            "type": "integer"
          },
          "inventory_value": {
            "title": "Inventory Value",
            "type": "number"
          },
          "low_stock": {
            "title": "Low Stock",
            "type": "integer"
          },
          "out_of_stock": {
            "title": "Out Of Stock",
            "type": "integer"
          },
          "total_units": {
            "title": "Total Units",
            "type": "integer"
          }
        },
        "required": [
          "active_products",
          "total_units",
          "inventory_value",
          "low_stock",
          "out_of_stock"
        ],
        "title": "InventoryStats",
        "type": "object"
      },
...
      "OrderCounts": {
        "properties": {
          "cancelled": {
            "title": "Cancelled",
            "type": "integer"
          },
          "completed": {
            "title": "Completed",
            "type": "integer"
          },
          "pending_approval": {
            "title": "Pending Approval",
            "type": "integer"
          },
          "rejected": {
            "title": "Rejected",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "pending_approval",
          "completed",
          "rejected",
          "cancelled"
        ],
        "title": "OrderCounts",
        "type": "object"
      },
...
      "Page_CustomerOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CustomerOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[CustomerOut]",
        "type": "object"
      },
...
      "SalesSummary": {
        "properties": {
          "average_order_value": {
            "title": "Average Order Value",
            "type": "number"
          },
          "completed_orders": {
            "title": "Completed Orders",
            "type": "integer"
          },
          "revenue_this_month": {
            "title": "Revenue This Month",
            "type": "number"
          },
          "revenue_today": {
            "title": "Revenue Today",
            "type": "number"
          },
          "total_revenue": {
            "title": "Total Revenue",
            "type": "number"
          }
        },
        "required": [
          "total_revenue",
          "revenue_this_month",
          "revenue_today",
          "completed_orders",
          "average_order_value"
        ],
        "title": "SalesSummary",
        "type": "object"
      },
      "SalesTrendPoint": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "orders": {
            "title": "Orders",
            "type": "integer"
          },
          "revenue": {
            "title": "Revenue",
            "type": "number"
          }
        },
        "required": [
          "date",
          "revenue",
          "orders"
        ],
        "title": "SalesTrendPoint",
        "type": "object"
      },
...
      "TopProduct": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "product_id": {
            "title": "Product Id",
            "type": "integer"
          },
          "quantity_sold": {
            "title": "Quantity Sold",
            "type": "integer"
          },
          "revenue": {
            "title": "Revenue",
            "type": "number"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          }
        },
        "required": [
          "product_id",
          "sku",
          "name",
          "quantity_sold",
          "revenue"
        ],
        "title": "TopProduct",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "The access_token returned by /auth/otp/verify",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Products, customers, sales orders, inventory and a manager approval workflow.",
    "title": "Sales & Inventory Management System",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/customers": {
      "get": {
        "operationId": "list_customers_api_v1_customers_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "is_active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "description": "1-based page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-based page number",
              "maximum": 10000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CustomerOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Customers",
        "tags": [
          "Customers"
        ]
      },
      "post": {
        "operationId": "create_customer_api_v1_customers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Customer",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/v1/customers/{customer_id}": {
      "delete": {
        "operationId": "deactivate_customer_api_v1_customers__customer_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "title": "Customer Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Deactivate a customer (soft delete)",
        "tags": [
          "Customers"
        ]
      },
      "get": {
        "operationId": "get_customer_api_v1_customers__customer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "title": "Customer Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Customer",
        "tags": [
          "Customers"
        ]
      },
      "patch": {
        "operationId": "update_customer_api_v1_customers__customer_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "title": "Customer Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Customer",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/v1/dashboard/sales-trend": {
      "get": {
        "operationId": "sales_trend_api_v1_dashboard_sales_trend_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 730,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SalesTrendPoint"
                  },
                  "title": "Response Sales Trend Api V1 Dashboard Sales Trend Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sales Trend",
        "tags": [
          "Dashboard & Settings"
        ]
      }
    },
    "/api/v1/dashboard/summary": {
      "get": {
        "operationId": "summary_api_v1_dashboard_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Summary",
        "tags": [
          "Dashboard & Settings"
        ]
      }
    },
    "/api/v1/dashboard/top-products": {
      "get": {
        "operationId": "top_products_api_v1_dashboard_top_products_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Only the last N days",
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 730,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only the last N days",
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TopProduct"
                  },
                  "title": "Response Top Products Api V1 Dashboard Top Products Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Top Products",
        "tags": [
          "Dashboard & Settings"
        ]
      }
    },
...
    "/api/v1/settings": {
      "get": {
        "operationId": "get_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingsOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Settings",
        "tags": [
          "Dashboard & Settings"
        ]
      },
      "put": {
        "operationId": "update_settings_api_v1_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Settings",
        "tags": [
          "Dashboard & Settings"
        ]
      }
    },
...
  }
}